requirements.txt
google-generativeai>=0.3.0
tqdm>=4.65.0 